Remove comments
