--- src/components/navigation-tabs/navigation-tabs.style.css
@@ -1,3 +1,0 @@
-button.bx--tabs--scrollable__nav-link.bx--tabs--scrollable.bx--tabs--scrollable__nav-link {
-  width: 20rem
-}
